# server/specs/openapi.yaml
openapi: 3.1.0
info:
  title: LoseIt API
  version: "0.1.0"
  license:
    name: MIT OR Apache-2.0
    identifier: MIT OR Apache-2.0
  description: |
    HTTP API for the LoseIt server. All routes (except `/health`) require a
    bearer token in the `Authorization` header. Token validation is delegated
    to a pluggable authenticator; the first identity seen for a given
    `(issuer, external_id)` pair is auto-provisioned into a local user row.

    All paths are served under the `/api/v1` prefix.

    Decimal-valued fields (weights, nutrition, quantities) are transported as
    JSON numbers but produced/consumed by the server as arbitrary-precision
    decimals — clients should avoid binary float round-trips for fields like
    `weight_kg`, `grams`, and `quantity`.

    ## v1 pre-release wire-shape changes

    The following endpoints returned bare JSON arrays in early drafts and
    now return paginated envelopes (`{ results, total, limit, offset }`):

    * `GET /log`
    * `GET /weights`

    Both endpoints' `from` and `to` query parameters were previously
    required and are now optional. With no parameters, the response is
    the 100 most-recent entries across the full history.

    The `GET /foods/search` response schema migrated from the bespoke
    `FoodSearchPage` shape to the unified `PaginatedFoodSearchHits`
    envelope; field names are unchanged.
servers:
  - url: "https://{host}/api/v1"
...
    variables:
      host:
        default: api.loseit.invalid
...
security:
  - bearerAuth: []
...
paths:
...
  /weights:
    post:
      tags: [Weights]
      operationId: create_weight
      summary: Record a weight entry
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WeightCreate" }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Weight" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
    get:
      tags: [Weights]
      operationId: list_weights
      summary: List weight entries
      description: |
        Lists the caller's weight entries, newest recorded-day first.

        All four query parameters are optional. With no parameters, the
        response is the 100 most-recent entries across the full history.
        Default `limit` is 100; server silently clamps requests above 500.
      parameters:
        - in: query
          name: from
          schema: { type: string, format: date }
          description: Inclusive lower bound on `recorded_on`.
        - in: query
          name: to
          schema: { type: string, format: date }
          description: Inclusive upper bound on `recorded_on`.
        - in: query
          name: limit
          schema: { type: integer, format: int64, minimum: 0 }
          description: Page size; 0 or omitted → 100.
        - in: query
          name: offset
          schema: { type: integer, format: int64, minimum: 0 }
          description: Page offset, starting at 0.
      responses:
        "200":
          description: Page of weight entries.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedWeights" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /weights/{id}:
    delete:
      tags: [Weights]
      operationId: delete_weight
      summary: Delete a weight entry
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "204": { description: Deleted. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /log:
    post:
      tags: [Log]
      operationId: create_log_entry
      summary: Record a food log entry
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LogCreate" }
      responses:
        "201":
          description: Created. The response includes a frozen nutrition snapshot.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LogEntry" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    get:
      tags: [Log]
      operationId: list_log_entries
      summary: List log entries
      description: |
        Lists the caller's log entries, newest consumed-day first.

        All four query parameters are optional. With no parameters, the
        response is the 100 most-recent entries across the full history.
        `from`/`to` are both inclusive bounds on `consumed_on`. Default
        `limit` is 100; server silently clamps requests above 500.
      parameters:
        - in: query
          name: from
          schema: { type: string, format: date }
          description: Inclusive lower bound on `consumed_on`.
        - in: query
          name: to
          schema: { type: string, format: date }
          description: Inclusive upper bound on `consumed_on`.
        - in: query
          name: limit
          schema: { type: integer, format: int64, minimum: 0 }
          description: Page size; 0 or omitted → 100.
        - in: query
          name: offset
          schema: { type: integer, format: int64, minimum: 0 }
          description: Page offset, starting at 0.
      responses:
        "200":
          description: Page of log entries.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedLogEntries" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /log/quick_add:
    post:
      tags: [Log]
      operationId: quick_add_log_entry
      summary: Log raw calories without choosing a food
      description: |
        Creates a log entry against a per-user sentinel "Quick Add" food
        owned by the caller. Only `calories_kcal` carries a value; all
        macronutrient fields on the returned snapshot are null. The
        sentinel food is hidden from `/foods/search`, `/foods/mine`,
        `/foods/recent`, and `/foods/frequent`. It is auto-provisioned on
        first use.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/QuickAddBody" }
      responses:
        "201":
          description: Created. The response includes a frozen snapshot with macros null.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LogEntry" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /log/copy:
    post:
      tags: [Log]
      operationId: copy_log_day
      summary: Copy a day's log entries to another day
      description: |
        Re-snapshots every entry in `from_date` (optionally filtered by
        `meal`) onto `to_date`. Snapshots are recomputed from the *current*
        food, so edits to a custom food between the two dates are
        reflected in the copied entries.

        Entries whose food is no longer visible to the caller, or whose
        serving has been deleted, are silently skipped. `from_date ==
        to_date` and `from_date > to_date` (backward copy) are both
        permitted. If `to_date` already has entries, they coexist with the
        new ones — nothing is replaced.

        The response is wrapped (`{ copied: [...] }`) rather than a bare
        array so future fields (e.g. `skipped`) can be added without
        breaking the wire shape.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CopyDayBody" }
      responses:
        "201":
          description: Created. Body contains the inserted entries, in input order.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CopyDayResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /log/{id}:
    patch:
      tags: [Log]
      operationId: update_log_entry
      summary: Update a log entry
      description: |
        `food_id` is immutable — sending one is a `400`. A `null` `note`
        clears the existing value; omitting the field leaves it unchanged.
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LogPatch" }
      responses:
        "200":
          description: Updated entry.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LogEntry" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Log]
      operationId: delete_log_entry
      summary: Delete a log entry
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "204": { description: Deleted. }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /days/{date}/summary:
    get:
      tags: [Days]
      operationId: get_day_summary
      summary: Per-day rollup of logged nutrition
      parameters:
        - in: path
          name: date
          required: true
          schema: { type: string, format: date }
          description: Calendar date to summarize.
      responses:
        "200":
          description: Day summary.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DaySummary" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Bearer token validated by the configured authenticator. In
        development a static `DEV_AUTH_BYPASS` token may be used; in
        production tokens are expected to be JWTs validated against a JWKS
        endpoint.
  parameters:
    PathId:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
      description: Resource identifier.
  responses:
    BadRequest:
      description: Request was malformed or failed domain validation.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing or invalid credentials.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Authenticated but not allowed to perform this action.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource does not exist or is not visible to the caller.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: Domain invariant prevents the operation.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable, machine-readable error identifier.
          examples:
            - bad_request
            - unauthorized
            - forbidden
            - not_found
            - conflict
            - upstream_error
            - internal_error
        message:
          type: string
    Decimal:
      description: Arbitrary-precision decimal serialized as a JSON number.
      type: number
    NullableDecimal:
      description: Arbitrary-precision decimal or `null`.
      type: [number, "null"]
...
    Meal:
      type: string
      enum: [breakfast, lunch, dinner, snack]
...
    Weight:
      type: object
      required: [id, recorded_on, weight_kg, created_at]
      properties:
        id: { type: string, format: uuid }
        recorded_on: { type: string, format: date }
        recorded_at_local:
          type: [string, "null"]
          description: Local time-of-day component (HH:MM:SS), if known.
        weight_kg: { $ref: "#/components/schemas/Decimal" }
        note:
          type: [string, "null"]
        created_at: { type: string, format: date-time }
    WeightCreate:
      type: object
      required: [recorded_on, weight_kg]
      properties:
        recorded_on: { type: string, format: date }
        recorded_at_local:
          type: string
          description: Local time-of-day component (HH:MM:SS).
        weight_kg: { $ref: "#/components/schemas/Decimal" }
        note: { type: string }
    Goal:
      type: object
      required: [id, starts_on, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        starts_on: { type: string, format: date }
        ends_on:
          type: [string, "null"]
          format: date
        start_weight_kg:
          $ref: "#/components/schemas/NullableDecimal"
        target_weight_kg:
          $ref: "#/components/schemas/NullableDecimal"
        weekly_rate_kg:
          $ref: "#/components/schemas/NullableDecimal"
        daily_calorie_target:
          type: [integer, "null"]
          format: int32
        protein_g_target:
          $ref: "#/components/schemas/NullableDecimal"
        carbs_g_target:
          $ref: "#/components/schemas/NullableDecimal"
        fat_g_target:
          $ref: "#/components/schemas/NullableDecimal"
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    QuickAddBody:
      type: object
      required: [calories_kcal, meal, consumed_on]
      properties:
        calories_kcal:
          allOf:
            - $ref: "#/components/schemas/Decimal"
          description: Must be > 0 and < 9999.
          exclusiveMinimum: 0
          maximum: 9998
        meal: { $ref: "#/components/schemas/Meal" }
        consumed_on: { type: string, format: date }
        note: { type: string }
    CopyDayBody:
      type: object
      required: [from_date, to_date]
      properties:
        from_date: { type: string, format: date }
        to_date: { type: string, format: date }
        meal:
          oneOf:
            - $ref: "#/components/schemas/Meal"
            - type: "null"
          description: If present, only entries with this meal are copied.
    CopyDayResponse:
      type: object
      required: [copied]
      properties:
        copied:
          type: array
          items: { $ref: "#/components/schemas/LogEntry" }
...
    Paginated:
      type: object
      required: [results, total, limit, offset]
      properties:
        results:
          type: array
          description: Replaced by the concrete schema in allOf compositions.
          items: {}
        total:
          type: integer
          format: int64
          description: Total matching rows across all pages.
        limit:
          type: integer
          format: int64
        offset:
          type: integer
          format: int64
...
    PaginatedLogEntries:
      allOf:
        - $ref: "#/components/schemas/Paginated"
        - type: object
          properties:
            results:
              type: array
              items: { $ref: "#/components/schemas/LogEntry" }
    PaginatedWeights:
      allOf:
        - $ref: "#/components/schemas/Paginated"
        - type: object
          properties:
            results:
              type: array
              items: { $ref: "#/components/schemas/Weight" }
    LogEntry:
      type: object
      description: |
        A logged consumption event with a frozen nutrition snapshot. The
        nutrition fields are flattened onto the entry rather than nested.
      required:
        - id
        - food_id
        - consumed_on
        - meal
        - quantity
        - grams_total
        - calories_kcal
        - created_at
        - updated_at
      properties:
        id: { type: string, format: uuid }
        food_id: { type: string, format: uuid }
        serving_id:
          type: [string, "null"]
          format: uuid
        consumed_on: { type: string, format: date }
        meal: { $ref: "#/components/schemas/Meal" }
        quantity: { $ref: "#/components/schemas/Decimal" }
        grams_total: { $ref: "#/components/schemas/Decimal" }
        calories_kcal: { $ref: "#/components/schemas/Decimal" }
        protein_g:
          $ref: "#/components/schemas/NullableDecimal"
        carbs_g:
          $ref: "#/components/schemas/NullableDecimal"
        fat_g:
          $ref: "#/components/schemas/NullableDecimal"
        fiber_g:
          $ref: "#/components/schemas/NullableDecimal"
        sugar_g:
          $ref: "#/components/schemas/NullableDecimal"
        sodium_mg:
          $ref: "#/components/schemas/NullableDecimal"
        saturated_fat_g:
          $ref: "#/components/schemas/NullableDecimal"
        note:
          type: [string, "null"]
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    LogCreate:
      type: object
      required: [food_id, serving_id, consumed_on, meal, quantity]
      properties:
        food_id: { type: string, format: uuid }
        serving_id: { type: string, format: uuid }
        consumed_on: { type: string, format: date }
        meal: { $ref: "#/components/schemas/Meal" }
        quantity: { $ref: "#/components/schemas/Decimal" }
        note: { type: string }
    LogPatch:
      type: object
      description: |
        Partial update. Sending `food_id` returns `400`. Sending `"note":
        null` clears the value; omitting the key leaves it untouched.
      properties:
        serving_id: { type: string, format: uuid }
        consumed_on: { type: string, format: date }
        meal: { $ref: "#/components/schemas/Meal" }
        quantity: { $ref: "#/components/schemas/Decimal" }
        note:
          type: [string, "null"]
    NutritionTotal:
      type: object
      required: [calories_kcal]
      properties:
        calories_kcal: { $ref: "#/components/schemas/Decimal" }
        protein_g:
          $ref: "#/components/schemas/NullableDecimal"
        carbs_g:
          $ref: "#/components/schemas/NullableDecimal"
        fat_g:
          $ref: "#/components/schemas/NullableDecimal"
        fiber_g:
          $ref: "#/components/schemas/NullableDecimal"
        sugar_g:
          $ref: "#/components/schemas/NullableDecimal"
        sodium_mg:
          $ref: "#/components/schemas/NullableDecimal"
        saturated_fat_g:
          $ref: "#/components/schemas/NullableDecimal"
    MealSubtotal:
      type: object
      required: [meal, calories_kcal, protein_g, carbs_g, fat_g, entry_count]
      properties:
        meal: { $ref: "#/components/schemas/Meal" }
        calories_kcal: { $ref: "#/components/schemas/Decimal" }
        protein_g: { $ref: "#/components/schemas/Decimal" }
        carbs_g: { $ref: "#/components/schemas/Decimal" }
        fat_g: { $ref: "#/components/schemas/Decimal" }
        entry_count: { type: integer, format: int32, minimum: 0 }
    DaySummary:
      type: object
      required: [date, total, by_meal]
      properties:
        date: { type: string, format: date }
        total: { $ref: "#/components/schemas/NutritionTotal" }
        by_meal:
          type: array
          description: One subtotal per meal in canonical order; empty meals included.
          items: { $ref: "#/components/schemas/MealSubtotal" }
        active_goal:
          oneOf:
            - $ref: "#/components/schemas/Goal"
            - type: "null"